+++ bindings/java/src/main/rust/lib.rs
@@ -80,6 +80,8 @@ fn build_inliner(
     let cache_size = env.get_int_field(&cfg, "cacheSize")?;
     let preallocate_node_capacity = env.get_int_field(&cfg, "preallocateNodeCapacity")?;
     let remove_inlined_selectors = env.get_bool_field(&cfg, "removeInlinedSelectors")?;
+    let apply_width_attributes = env.get_bool_field(&cfg, "applyWidthAttributes")?;
+    let apply_height_attributes = env.get_bool_field(&cfg, "applyHeightAttributes")?;
 
     let extra_css = env.get_string_field_opt(&cfg, "extraCss")?;
     let base_url = env.get_string_field_opt(&cfg, "baseUrl")?;
@@ -92,7 +94,9 @@ fn build_inliner(
         .load_remote_stylesheets(load_remote_stylesheets)
         .extra_css(extra_css.map(Cow::Owned))
         .preallocate_node_capacity(preallocate_node_capacity as usize)
-        .remove_inlined_selectors(remove_inlined_selectors);
+        .remove_inlined_selectors(remove_inlined_selectors)
+        .apply_width_attributes(apply_width_attributes)
+        .apply_height_attributes(apply_height_attributes);
 
     if let Some(url) = base_url {
         match css_inline::Url::parse(&url) {